+++ inc/customizer/assets/js/woo-typography-customize-preview.js
@@ -3,7 +3,7 @@
  *
  * @version 1.0.0
  * 
- * @since 1.8.8
+ * @since 2.0
  */
 ( function( $ ) {